Work is progress
This is a work in progress for the Slave/Master I2C protocol, it is not finished and hence there is a lot of code that is commented out. The programme compiles fine on my end at least.

A lot of alarm logic has also been implemented, however I was not able to build the physical board and test the code on it, but as previously mentioed it does compile.